feat(spi): Remove the production of a TopologyConfigBuildItem
Indeed the deployment of the Quarkus extension was producing a TopologyConfigBuildItem that could then be reused in the processor of another Quarkus extension to get access to the resolved topic mappings.
However, using the KStreamProcessorConfigRuntime in a normal @BuildStep of the processor means that the "runtime" configuration becomes build time.
Indeed if added as expected the ConfigPhase.RUN_TIME to the ConfigRoot annotation, the processor's tests start failing.

Solution: entirely remove the TopologyConfigBuildItem and moving the mapping building classes to `spi`.

BREAKING CHANGE: The TopologyConfigBuildItem is no more available.
Instead we provide SinkToTopicMappingBuilder and SourceToTopicsMappingBuilder with constructor expecting smallrye-config's Config object.

BREAKING CHANGE: Update the base versions in POMs to reflect the upcoming major release.