chore(release): Release 4.0.1
Changing version to 4.0.0-SNAPSHOT to reflect the update to Quarkus 3.15 and the creation of 3.8 with the fix of Decorators.
Here the Decorator fix is restored, as the flakiness can still be seen on [master](https://github.com/quarkiverse/quarkus-kafka-streams-processor/actions/runs/11296240001/job/31420666727).
The original PR was [here](https://github.com/quarkiverse/quarkus-kafka-streams-processor/pull/118).

The documentation is updated to only talk about ProcessorDecorator using the AbstractProcessorDecorator abstract class.