fix(impl): Support custom span creation in Processor
Probably a very narrow use case.
But it could happen.
Along the way of a [recent PR](https://github.com/quarkiverse/quarkus-kafka-streams-processor/pull/109) by @nreant1A trying to make sure:

1. Baggage is properly propagated
2. Custom span creationg is supported in process method

First added 1 QuarkusTest in impl.

The implementation was:

1. extracting the parent context (but not making it current, by which baggage were lost)
2. creating a "process" span and making that a current span
3. early on reinjecting the new span context ids in the message before it starts being processed, relying on the fact that output message are created by modification of the incoming one: `ping.withValue(...)`

The issue we have: baggage cannot be modified in the processor, and any other context modification in the processor is basically ignored (creating a custom span for instance).

Solution: move the injection of the contextual data (traceparent, tracestate, baggage) at message production by implementing a TracingProducerInterceptor.

The OpenTelemetryExtension for JUnit 5 has bean copy pasted so we can inject the W3CBaggagePropagator.
Indeed the original class is final, with private constructor.
A fix upstream will be done shortly.

Also the KStreamTopologyDriverTest started failing in tests related to tracing.
Actually in tests based on TopologyTestDriver, the ProducerOnSendInterceptors are not called.
The solution is to transform those tests methods in ones using the Kafka Devservices.