feat(docs): Restore doc about non CDI decorator and Quarkus 3.8
Also restore the corresponding integration test that was testing this use case.
Whilst doing that noticed that the QuarkusTest is becoming flaky again.
So reintroducing the `quarkus.test.flat-class-path=true` in `impl` so noone will be impacted by this flakiness anymore.

Also noting the `quarkus-3.8` branch did not have a build, whilst it holds the 2.0x branch which we will need to build in parallel.