feat(impl): Support interception of output record before serialization
Fixes #147.

Introduction of a new `OutputRecordInterceptor` interface to be implemented to be able to intercept outgoing messages before their payload is serialized.
Its aim is to replace the `ProducerOnSendInterceptor` that at the difference with `OutputRecordInterceptor` is called after the serialization of the payload.
Henceforth `ProducerOnSendInterceptor` is deprecated.

Version is bumped to 4.1 to reflect the new feature.
It is backwards-compatible.

The TracingProducerInterceptor has been rewritten as a TracingOutputRecordInterceptor, but kept for backwards-compatibility.
It is itself marked for removal as well.

Documentation updated accordingly.