feat(deps): Migrate to Quarkus 3.24 and Kafka 4.0
Main breaking change: the Kafka 2 API for Processor has been removed from Kafka Client 4.0.0.
It allows the library to clean up all backwards compatibility layers introduced to work both with Kafka 3 and 2 APIs.

Another change is the MockProcessorContext, that can no more be passed directly to a `StateStore.init` call, has to be a `StateStoreContext`, fortunately easy to access through `MockProcessorContext.getStateStoreContext()`.

Because of the breaking change involved, setting version to 5.0.0-SNAPSHOT.