fix(docs): Add quarkus processor to spi
Currently main is not building because the configuration documentation generation is failing, complaining of missing javadoc.
Turns out the pb is we do not run the quarkus-extension-processor on the spi module, which contains the actual configuration interfaces.
Fortunately adding the quarkus-extension-processor on the `spi` module seems to solve the issue.

However it had a side effect on `KStreamsProcessorConfigGenerator`: the private classes implementing the config interfaces need now a default constructor and some incompatibility with lombok was found.
Consequently removed lombok annotations, `final` keywords on field declaration, and write getter/setters manually.