feat(impl): Add support for CloudEvents natively to the library
A new CloudEventContextHandler can be injected to get access to the cloud event metadata received in input, and a setter to modify the outgoing metadata.
It also proposes a customized CloudEventBuilder implementation that is doing some defaulting of the fields id, source and type, based on configuration or generated string.

The activation is done separately for input and output, but with all-or-nothing feature flags `kafkastreamsprocessor.input.is-cloud-event` and `kafkastreamsprocessor.output.is-cloud-event`.
The serializer/deserializer can be customized with the configuration Map `kafkastreamsprocessor.input.cloud-event-deserializer-config` and `kafkastreamsprocessor.output.cloud-event-serializer-config`.

For this feature, the ProducerOnSendInterceptor is being reintroduced, as we need to wrap the payload in a cloud event only after the message payload has been serialized to byte[] by the Topology.
Its deprecation is this removed, and its documentation polished.
A documentation section has been added to cover all about `CloudEvent` data format.

Fixes #219.