Migrate to Quarkus 3.30 and Protobuf Java 4
Fixes #224.

Quarkus 3.27 LTS depends on protobuf-java 3.x whilst the latest Quarkus 3.30.3 depends on protobuf-java 4.x.
Fortunately the library is backwards-compatible, and we make it support the new GeneratedMessage base class type.
The backwards- and forwards-compatibility is tested with 2 new protobuf-binding and 2 corresponding simple processor samples with:

* generation of Java classes from a proto file with proto3 syntax and the protoc 3.x tool
* generation of Java classes from a proto file with no syntax and edition 2024 and the protoc 4.x tool

But eventually, it is still an update of major version of a dependency, so we mark it with a major version change of the extension itself.